+++ aldryn_forms/middleware/handle_post.py
@@ -7,7 +7,7 @@ from django.utils.deprecation import MiddlewareMixin
 from aldryn_forms.constants import ALDRYN_FORMS_POST_IDENT_NAME
 from aldryn_forms.forms import FormSubmissionBaseForm
 from aldryn_forms.models import FormPlugin
-from aldryn_forms.utils import get_plugin_tree
+from aldryn_forms.utils import get_form_anchor, get_plugin_tree
 
 
 class HandleHttpPost(MiddlewareMixin):
@@ -62,6 +62,9 @@ def get_response(
     if data["status"] == "SUCCESS":
         form_plugin_instance = form_plugin.get_plugin_instance()[1]
         success_url = form_plugin_instance.get_success_url(instance=form_plugin, post_ident=data.get("post_ident"))
+        if form_plugin.message_on_form:
+            success_url += "#" + get_form_anchor(form_plugin.pk)
+            request.session[get_form_anchor(form_plugin.pk)] = True
         if success_url:
             return HttpResponseRedirect(success_url)