+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 Unreleased
 ==========
 
+- Unification of the message "Please wait. Submitting form..." in form.js.
 - Disable sending webhook in DEBUG mode. Added webhook testing via Constance.
 
 8.9.3 (2026-06-01)